requirements.txt
aiogram==3.22.0
python-dotenv==1.0.1
groq==0.13.1
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
openpyxl==3.1.5
tzdata